stringToPolynomial method
static stringToPolynomial(...) added to Polynomial to convert a given string (in the provided format) to the Polynomial DataType.

Adapted hasDegree() method in Polynomial to handle 1-sized lists consisting of only zeros as well.

Fixed a bug in Polynomial's toString() method where the order of each element was one higher than what it should be.

Changed all calls to poly.isEmpty() to !hasDegree()

Adapted InputReader to work with the new input-fields. This has not yet been tested.