Addition and Multiplication Tables + FF-update + Polynomial.copy() BugFix
Fixed a bug in Polynomial.copy() where a reference to the internal LinkedList was also copied over. The LinkedList is now correctly copied, and thus no reference is preserved.

Changed the generateElements() method in FiniteField to now generate elements with lowest degree to highest degree, instead of the other way around.
Changed FiniteField.contains(..) to consider congruent Polynomials as well.

Added generateAdditionTable() and generateMultiplicationTable() methods to FiniteField.