FiniteFieldArithmetic.isPrimitive(...) + more methods + Bugfixes
Added FiniteFieldArithmetic.isPrimitive(Polynomial, FiniteField), which tests if a given Polynomial is a primitive element of a given Finite Field.
Added Arithmetic.getPrimeDivisors(int), which (inefficiently) finds the prime divisors of a given integer.
Added FiniteField.getOrder()
Added FiniteFieldArithmetic.powerize(Polynomial, FiniteField f, int), which computes Polynomial^int (by calling multiplication several times)
Added Polynomial.copy(int), which copies the Polynomial with a given modulus.

Fixed a Bug in Arithmetic where isPrimitive(..) was called on the polyMod (which always resulted in 'false') instead of the given Polynomial.

Fixed a bug in Division where, if the divisor had a larger degree than the number to divide, and the number to divide had coefficients larger than the modulus, the modulus was not applied to the number to divide's coefficients.